separated out pip component and updated styling
