+++ src/components/CameraFeed.tsx
@@ -12,6 +12,7 @@ import {
   AlertDialogTitle,
 } from "./ui/alert-dialog";
 import { CameraOff, Maximize2 } from "lucide-react";
+import { usePictureInPicture } from "@/lib/usePictureInPicture";
 
 interface CameraFeedProps {
   onPermissionGranted?: () => void;
@@ -44,12 +45,19 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
       (ref as React.RefObject<HTMLVideoElement>) || localVideoRef;
     const [hasPermission, setHasPermission] = useState<boolean | null>(null);
     const [showPermissionDialog, setShowPermissionDialog] = useState(true);
-    const [pipWindow, setPipWindow] = useState<Window | null>(null);
     const [isRunning, setIsRunning] = useState(false);
     const [remainingTime, setRemainingTime] = useState(0);
     const [taskName, setTaskName] = useState("");
     const [duration, setDuration] = useState(25 * 60); // Store duration in seconds
     const timerRef = useRef<NodeJS.Timeout>();
+
+    // Use the PiP hook
+    const { enterPiP, showEndMessage } = usePictureInPicture({
+      videoRef,
+      taskName,
+      remainingTime,
+      onSessionComplete,
+    });
 
     const startCamera = async () => {
       try {
@@ -68,215 +76,6 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
       }
     };
 
-    const showEndMessage = () => {
-      if (pipWindow) {
-        const timerDiv = pipWindow.document.querySelector(".pip-timer");
-        if (timerDiv) {
-          timerDiv.innerHTML = `
-            <div class="pip-end-message">
-              <svg class="checkmark" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 52 52">
-                <circle class="checkmark-circle" cx="26" cy="26" r="25" fill="none"/>
-                <path class="checkmark-check" fill="none" d="M14.1 27.2l7.1 7.2 16.7-16.8"/>
-              </svg>
-              <div class="pip-task-name">Session Complete 🎉</div>
-            </div>
-          `;
-        }
-      }
-    };
-
-    const enterPiP = async () => {
-      if (!("documentPictureInPicture" in window)) {
-        console.error("Picture-in-Picture API not supported");
-        return;
-      }
-
-      try {
-        // Close existing PiP window if it exists
-        if (pipWindow) {
-          pipWindow.close();
-        }
-
-        // @ts-ignore - TypeScript doesn't know about Document Picture-in-Picture API yet
-        const newPipWindow =
-          await window.documentPictureInPicture.requestWindow({
-            width: 400,
-            height: 300,
-          });
-
-        setPipWindow(newPipWindow);
-
-        // Create styles for the PiP window
-        const style = document.createElement("style");
-        style.textContent = `
-          body { 
-            margin: 0;
-            background: transparent;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            height: 100vh;
-            font-family: system-ui;
-          }
-          .pip-container {
-            width: 100%;
-            height: 100%;
-            position: relative;
-            background: rgba(0, 0, 0, 0.8);
-            border-radius: 8px;
-            overflow: hidden;
-          }
-          .pip-video {
-            width: 100%;
-            height: 100%;
-            object-fit: cover;
-            transform: rotateY(180deg);
-          }
-          .pip-timer {
-            position: absolute;
-            inset: 0;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            gap: 8px;
-            background: rgba(0, 0, 0, 0.3);
-          }
-          .pip-task-name {
-            background: rgb(23 23 23 / 0.5);
-    backdrop-filter: blur(4px);
-    -webkit-backdrop-filter: blur(4px);
-            padding: 0.5rem 1rem;
-            border-radius: 0.5rem;
-            font-size: 1rem;
-            color: white;
-          }
-          .pip-timer-text {
-            font-size: 2rem;
-            font-weight: bold;
-            color: white;
-          }
-          .pip-end-message {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            gap: 1rem;
-          }
-          .checkmark {
-            width: 56px;
-            height: 56px;
-            border-radius: 50%;
-            stroke-width: 2;
-            stroke: #fff;
-            stroke-miterlimit: 10;
-            animation: scale 0.3s ease-in-out;
-          }
-          .checkmark-circle {
-            stroke-dasharray: 166;
-            stroke-dashoffset: 166;
-            stroke-width: 2;
-            stroke: #fff;
-            fill: none;
-            animation: stroke 0.6s cubic-bezier(0.65, 0, 0.45, 1) forwards;
-          }
-          .checkmark-check {
-            transform-origin: 50% 50%;
-            stroke-dasharray: 48;
-            stroke-dashoffset: 48;
-            animation: stroke 0.3s cubic-bezier(0.65, 0, 0.45, 1) 0.8s forwards;
-          }
-          @keyframes stroke {
-            100% {
-              stroke-dashoffset: 0;
-            }
-          }
-          @keyframes scale {
-            0%, 100% {
-              transform: none;
-            }
-            50% {
-              transform: scale3d(1.1, 1.1, 1);
-            }
-          }
-        `;
-        newPipWindow.document.head.appendChild(style);
-
-        const container = document.createElement("div");
-        container.className = "pip-container";
-
-        const video = document.createElement("video");
-        video.className = "pip-video";
-        video.autoplay = true;
-        video.playsInline = true;
-        video.muted = true;
-        if (videoRef.current?.srcObject) {
-          video.srcObject = videoRef.current.srcObject;
-        }
-        container.appendChild(video);
-
-        // Timer container
-        const timerDiv = document.createElement("div");
-        timerDiv.className = "pip-timer";
-
-        const taskNameDiv = document.createElement("div");
-        taskNameDiv.className = "pip-task-name";
-        taskNameDiv.textContent = taskName;
-        timerDiv.appendChild(taskNameDiv);
-
-        const timerText = document.createElement("span");
-        timerText.className = "pip-timer-text";
-        timerDiv.appendChild(timerText);
-        container.appendChild(timerDiv);
-
-        newPipWindow.document.body.appendChild(container);
-
-        // Handle window closing
-        const handleUnload = () => {
-          setPipWindow(null);
-        };
-
-        newPipWindow.addEventListener("unload", handleUnload);
-        newPipWindow.addEventListener("beforeunload", handleUnload);
-
-        // Handle page visibility changes
-        const handleVisibilityChange = () => {
-          if (document.hidden && newPipWindow) {
-            video.srcObject = videoRef.current?.srcObject || null;
-          }
-        };
-
-        document.addEventListener("visibilitychange", handleVisibilityChange);
-
-        // Cleanup when component unmounts
-        return () => {
-          document.removeEventListener(
-            "visibilitychange",
-            handleVisibilityChange,
-          );
-          if (newPipWindow) {
-            newPipWindow.removeEventListener("unload", handleUnload);
-            newPipWindow.removeEventListener("beforeunload", handleUnload);
-            newPipWindow.close();
-          }
-        };
-      } catch (err) {
-        console.error("Error opening Picture-in-Picture window:", err);
-        setPipWindow(null);
-      }
-    };
-
-    // Update PiP window timer
-    useEffect(() => {
-      if (pipWindow) {
-        const timerText = pipWindow.document.querySelector(".pip-timer-text");
-        if (timerText) {
-          timerText.textContent = `${Math.floor(remainingTime / 60)}:${String(
-            Math.floor(remainingTime % 60),
-          ).padStart(2, "0")}`;
-        }
-      }
-    }, [remainingTime, pipWindow]);
-
     // Cleanup timer on unmount
     useEffect(() => {
       return () => {
@@ -293,7 +92,7 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
         showEndMessage();
         onSessionComplete();
       }
-    }, [remainingTime, isRunning]);
+    }, [remainingTime, isRunning, showEndMessage, onSessionComplete]);
 
     useEffect(() => {
       // Check if browser supports getUserMedia
@@ -321,17 +120,14 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
           ).getTracks();
           tracks.forEach((track) => track.stop());
         }
-        if (pipWindow) {
-          pipWindow.close();
-        }
       };
     }, []);
 
     return (
       <Card className="w-full h-full bg-background relative border-none inner-stroke-black-10-sm overflow-hidden shadow-sm rounded-lg">
         {hasPermission ? (
-          <div className="relative w-full h-full  overflow-hidden">
-            {/* 1) No blur here on the camera feed */}
+          <div className="relative w-full h-full overflow-hidden">
+            {/* Camera feed */}
             <video
               ref={videoRef}
               autoPlay
@@ -452,7 +248,7 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
               {/* Running state: show timer */}
               {isRunning && (
                 <div className="absolute inset-0 flex items-center justify-center w-full">
-                  <div className="px-4 sm:px-6 py-3 sm:py-4 rounded-xl bg-gradient-to-b from-neutral-700/50 via-neutral-900/50 to-neutral-900/50  shadow-sm backdrop-blur-md border-none  inner-stroke-white-10-sm">
+                  <div className="px-4 sm:px-6 py-3 sm:py-4 rounded-xl bg-gradient-to-b from-neutral-700/50 via-neutral-900/50 to-neutral-900/50 shadow-sm backdrop-blur-md border-none inner-stroke-white-10-sm">
                     <div className="text-2xl sm:text-3xl font-bold text-white/90 text-center">
                       {Math.floor(remainingTime / 60)}:
                       {String(Math.floor(remainingTime % 60)).padStart(2, "0")}
@@ -482,7 +278,7 @@ const CameraFeed = React.forwardRef<HTMLVideoElement, CameraFeedProps>(
               <Button
                 variant="secondary"
                 className="bg-white/60 before:absolute before:inset-0 before:bg-gradient-to-b before:from-transparent before:to-black/20 before:rounded-full text-black/75 backdrop-blur-md flex items-center gap-2 rounded-full inner-stroke-white-20-sm"
-                onClick={enterPiP}
+                onClick={() => enterPiP({ width: 400, height: 300 })}
               >
                 <Maximize2 className="h-4 w-4" />
                 Open floating window
